refactor search and category pages using mixins, refactor types
