+++ apps/storefront-e2e/src/support/cy-actions/login.actions.ts
@@ -1,6 +1,6 @@
 import { GlueAPI } from '../apis/glue.api';
 import { LoginPage } from '../page-objects/login.page';
-import { Customer } from '../types/user.type';
+import { Customer } from '../types/domain.types';
 
 export {};
