+++ apps/storefront-e2e/src/support/mixins/page-with-product-sorting.mixin.ts
@@ -1,0 +1,22 @@
+import { ProductSortingFragment } from '../page-fragments/products-sorting.fragment';
+import { E2EPage } from '../page-objects/abstract.page';
+import { Constructor } from '../types/utils.types';
+
+export interface IPageWithProductSorting {
+  getProductSorting(): ProductSortingFragment;
+}
+
+export function WithProductSorting<TPage extends Constructor<E2EPage>>(
+  Page: TPage
+) {
+  return class PageWithProductSorting
+    extends Page
+    implements IPageWithProductSorting
+  {
+    constructor(...args: any[]) {
+      super(...args);
+    }
+
+    getProductSorting = () => new ProductSortingFragment();
+  };
+}
