+++ apps/storefront-e2e/src/support/page-objects/category.page.ts
@@ -1,17 +1,10 @@
-import {
-  IPageWithFacets,
-  IPageWithProductList,
-  IPageWithProductSorting,
-} from '../interfaces/page-interfaces';
-import { FacetsFragment } from '../page-fragments/facets.fragment';
-import { ProductSortingFragment } from '../page-fragments/products-sorting.fragment';
-import { Category } from '../types/category.type';
+import { WithFacets } from '../mixins/page-with-facets.mixin';
+import { WithProductList } from '../mixins/page-with-product-list.mixin';
+import { WithProductSorting } from '../mixins/page-with-product-sorting.mixin';
+import { Category } from '../types/domain.types';
 import { AbstractSFPage } from './abstract.page';
 
-export class CategoryPage
-  extends AbstractSFPage
-  implements IPageWithProductList, IPageWithFacets, IPageWithProductSorting
-{
+export class CategoryPageBase extends AbstractSFPage {
   url = '/category/';
   categoryId: string;
 
@@ -22,27 +15,13 @@ export class CategoryPage
       this.categoryId = categoryData.id;
       this.url += categoryData.id;
     }
-
-    this.initSearchInterceptor();
-  }
-
-  initSearchInterceptor(): void {
-    cy.intercept('/catalog-search?**').as('catalogSearch');
-  }
-
-  waitForSearchRequest(): void {
-    cy.wait('@catalogSearch');
-    // wait till product cards are re-renreded after search
-    // eslint-disable-next-line cypress/no-unnecessary-waiting
-    cy.wait(500);
   }
 
   waitForLoaded(): void {
-    this.waitForSearchRequest();
+    //
   }
-
-  getProductSorting = () => new ProductSortingFragment();
-  getFacets = () => new FacetsFragment();
-  getProductCards = () => cy.get('oryx-product-card');
-  getProductHeadings = () => this.getProductCards().find('oryx-heading');
 }
+
+export const CategoryPage = WithProductSorting(
+  WithProductList(WithFacets(CategoryPageBase))
+);
