+++ apps/storefront-e2e/src/support/test-data/search-products.ts
@@ -1,0 +1,39 @@
+export const sortingTestData = {
+  default: ['cable-hdmi-1-1', 'cable-vga-1-1', '216_123', '217_123', '215_123'],
+  rating: ['216_123', '217_123', '215_123', 'cable-hdmi-1-1', 'cable-vga-1-1'],
+  name_asc: [
+    '217_123',
+    '215_123',
+    'cable-hdmi-1-1',
+    '216_123',
+    'cable-vga-1-1',
+  ],
+  name_desc: [
+    'cable-vga-1-1',
+    '216_123',
+    'cable-hdmi-1-1',
+    '215_123',
+    '217_123',
+  ],
+  price_asc: [
+    'cable-hdmi-1-1',
+    'cable-vga-1-1',
+    '215_123',
+    '216_123',
+    '217_123',
+  ],
+  price_desc: [
+    '217_123',
+    '216_123',
+    '215_123',
+    'cable-hdmi-1-1',
+    'cable-vga-1-1',
+  ],
+  popularity: [
+    'cable-vga-1-1',
+    'cable-hdmi-1-1',
+    '215_123',
+    '216_123',
+    '217_123',
+  ],
+};
