+++ apps/storefront-e2e/src/support/test-data/storages/product.storage.ts
@@ -1,4 +1,4 @@
-import { Product } from '../../types/product.type';
+import { Product } from '../../types/domain.types';
 
 const products: Product[] = [
   {
