+++ apps/storefront-e2e/src/support/test-data/storages/search-parameters.storage.ts
@@ -1,4 +1,4 @@
-import { SearchParameters } from '../../types/search.type';
+import { SearchParameters } from '../../types/domain.types';
 
 const parameters: SearchParameters[] = [
   {
