--- apps/storefront-e2e/src/support/types/category.type.ts
@@ -1,3 +1,0 @@
-export type Category = {
-  id: string;
-};
