--- apps/storefront-e2e/src/support/types/product.type.ts
@@ -1,8 +1,0 @@
-export type Product = {
-  id: string;
-  title: string;
-  originalPrice: string;
-  currentPrice?: string;
-  currentPriceWith10pDiscount?: string;
-  previewImageURL?: string;
-};
