--- apps/storefront-e2e/src/support/types/search.type.ts
@@ -1,3 +1,0 @@
-export type SearchParameters = {
-  q: string;
-};
