--- apps/storefront-e2e/src/support/types/user.type.ts
@@ -1,6 +1,0 @@
-export type Customer = {
-  id?: string;
-  name: string;
-  email: string;
-  password: string;
-};
