+++ apps/storefront-e2e/src/support/types/utils.types.ts
@@ -1,0 +1,1 @@
+export type Constructor<T = object> = new (...args: any[]) => T;
